+++ CHANGELOG.md
@@ -17,10 +17,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased](https://github.com/o1-labs/o1js/compare/e1bac02...HEAD)
 
+### Added
+
+- API for recursively proving a ZkProgram method from within another https://github.com/o1-labs/o1js/pull/1931
+  - `program.proveRecursively.<methodName>(...args): Promise<PublicOutput>`
+  - This also works within the same program, as long as the return value is type-annotated
+
 ### Fixed
 
 - Compiling stuck in the browser for recursive zkprograms https://github.com/o1-labs/o1js/pull/1906
-
 - Error message in `rangeCheck16` gadget https://github.com/o1-labs/o1js/pull/1920
 
 ## [2.1.0](https://github.com/o1-labs/o1js/compare/b04520d...e1bac02) - 2024-11-13